pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "utils"
description = "Utilities to add dynamic information in GitHub profile."
readme = "README.md"
requires-python = "==3.11"
license = {text = "MIT"}
classifiers = [
    "Programming Language :: Python :: 3",
]
dependencies = [
    "requests==2.28.1",
    "cogapp==3.3.0",
    "python-dateutil==2.8.2",
    "loguru==0.6.0"
]

dynamic = ["version"]

[project.optional-dependencies]
qa = ["pre-commit==2.20.0"]
